Move instrumentCache to pipeline.go
